+++ src/main/java/com/rs/math/geometry/shape/Point.java
@@ -3,9 +3,9 @@ package com.rs.math.geometry.shape;
 import com.rs.math.geometry.Constants;
 
 public class Point {
-    public final float x;
-    public final float y;
-    public Point(float x, float y) {
+    public final double x;
+    public final double y;
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
@@ -25,7 +25,7 @@ public class Point {
     public static boolean same(Point p1, Point p2) {
         return same(p1, p2, Constants.EPSILON);
     }
-    public static boolean same(Point p1, Point p2, float epsilon) {
+    public static boolean same(Point p1, Point p2, double epsilon) {
         return Math.abs(p1.x - p2.x) < epsilon && Math.abs(p1.y - p2.y) < epsilon;
     }
 }
