+++ src/test/java/com/rs/math/geometry/func/ConvexHullTest.java
@@ -10,17 +10,17 @@ public class ConvexHullTest {
 
     @Test
     public void simpleConvexHull_JarvisMarch() throws Exception {
-        testConvexHull_JarvisMarch(new float[]{1, 1, 0, 0, 2, 2, 2, 0, 0, 2}, new float[]{0, 0, 0, 2, 2, 2, 2, 0});
-        testConvexHull_JarvisMarch(new float[]{1, 1, 0, 0, 3, 1, 2, 2, 2, 0, 0, 2}, new float[]{0, 0, 0, 2, 2, 2, 3, 1, 2, 0});
+        testConvexHull_JarvisMarch(new double[]{1, 1, 0, 0, 2, 2, 2, 0, 0, 2}, new double[]{0, 0, 0, 2, 2, 2, 2, 0});
+        testConvexHull_JarvisMarch(new double[]{1, 1, 0, 0, 3, 1, 2, 2, 2, 0, 0, 2}, new double[]{0, 0, 0, 2, 2, 2, 3, 1, 2, 0});
     }
 
     @Test
     public void complexConvexHull_JarvisMarch() throws Exception {
-        testConvexHull_JarvisMarch(new float[]{1, 1, 0, 0, 2, 2, 2, 0, 0, 2, 1, 2.00001f}, new float[]{0, 0, 0, 2, 2, 2, 2, 0});
-        testConvexHull_JarvisMarch(new float[]{0, 0, 1, 0, 0, 1, 1 / 3f, 2 / 3f, 2 / 3f, 1 / 3f}, new float[]{0, 0, 1, 0, 0, 1});
+        testConvexHull_JarvisMarch(new double[]{1, 1, 0, 0, 2, 2, 2, 0, 0, 2, 1, 2.00001f}, new double[]{0, 0, 0, 2, 2, 2, 2, 0});
+        testConvexHull_JarvisMarch(new double[]{0, 0, 1, 0, 0, 1, 1 / 3f, 2 / 3f, 2 / 3f, 1 / 3f}, new double[]{0, 0, 1, 0, 0, 1});
     }
 
-    private void testConvexHull_JarvisMarch(float[] in, float[] out) {
+    private void testConvexHull_JarvisMarch(double[] in, double[] out) {
         List<Point> points = new Polygon(in).points;
         Polygon polygon = ConvexHull.convexHull_JarvisMarch(points);
         List<Point> ref = new Polygon(out).points;
@@ -30,17 +30,17 @@ public class ConvexHullTest {
 
     @Test
     public void simpleConvexHull_AndrewMonotoneChain() throws Exception {
-        testConvexHull_AndrewMonotoneChain(new float[]{1, 1, 0, 0, 2, 2, 2, 0, 0, 2}, new float[]{0, 0, 0, 2, 2, 2, 2, 0});
-        testConvexHull_AndrewMonotoneChain(new float[]{1, 1, 0, 0, 3, 1, 2, 2, 2, 0, 0, 2}, new float[]{0, 0, 0, 2, 2, 2, 3, 1, 2, 0});
+        testConvexHull_AndrewMonotoneChain(new double[]{1, 1, 0, 0, 2, 2, 2, 0, 0, 2}, new double[]{0, 0, 0, 2, 2, 2, 2, 0});
+        testConvexHull_AndrewMonotoneChain(new double[]{1, 1, 0, 0, 3, 1, 2, 2, 2, 0, 0, 2}, new double[]{0, 0, 0, 2, 2, 2, 3, 1, 2, 0});
     }
 
     @Test
     public void complexConvexHull_AndrewMonotoneChain() throws Exception {
-        testConvexHull_AndrewMonotoneChain(new float[]{1, 1, 0, 0, 2, 2, 2, 0, 0, 2, 1, 2.00001f}, new float[]{0, 0, 0, 2, 2, 2, 2, 0});
-        testConvexHull_AndrewMonotoneChain(new float[]{0, 0, 1, 0, 0, 1, 1 / 3f, 2 / 3f, 2 / 3f, 1 / 3f}, new float[]{0, 0, 1, 0, 0, 1});
+        testConvexHull_AndrewMonotoneChain(new double[]{1, 1, 0, 0, 2, 2, 2, 0, 0, 2, 1, 2.00001f}, new double[]{0, 0, 0, 2, 2, 2, 2, 0});
+        testConvexHull_AndrewMonotoneChain(new double[]{0, 0, 1, 0, 0, 1, 1 / 3f, 2 / 3f, 2 / 3f, 1 / 3f}, new double[]{0, 0, 1, 0, 0, 1});
     }
 
-    private void testConvexHull_AndrewMonotoneChain(float[] in, float[] out) {
+    private void testConvexHull_AndrewMonotoneChain(double[] in, double[] out) {
         List<Point> points = new Polygon(in).points;
         Polygon polygon = ConvexHull.convexHull_AndrewMonotoneChain(points);
         List<Point> ref = new Polygon(out).points;
@@ -50,17 +50,17 @@ public class ConvexHullTest {
 
     @Test
     public void simpleConvexHull_misdake() throws Exception {
-        testConvexHull_misdake(new float[]{1, 1, 0, 0, 2, 2, 2, 0, 0, 2}, new float[]{0, 0, 0, 2, 2, 2, 2, 0});
-        testConvexHull_misdake(new float[]{1, 1, 0, 0, 3, 1, 2, 2, 2, 0, 0, 2}, new float[]{0, 0, 0, 2, 2, 2, 3, 1, 2, 0});
+        testConvexHull_misdake(new double[]{1, 1, 0, 0, 2, 2, 2, 0, 0, 2}, new double[]{0, 0, 0, 2, 2, 2, 2, 0});
+        testConvexHull_misdake(new double[]{1, 1, 0, 0, 3, 1, 2, 2, 2, 0, 0, 2}, new double[]{0, 0, 0, 2, 2, 2, 3, 1, 2, 0});
     }
 
     @Test
     public void complexConvexHull_misdake() throws Exception {
-        testConvexHull_misdake(new float[]{1, 1, 0, 0, 2, 2, 2, 0, 0, 2, 1, 2.00001f}, new float[]{0, 0, 0, 2, 2, 2, 2, 0});
-        testConvexHull_misdake(new float[]{0, 0, 1, 0, 0, 1, 1 / 3f, 2 / 3f, 2 / 3f, 1 / 3f}, new float[]{0, 0, 1, 0, 0, 1});
+        testConvexHull_misdake(new double[]{1, 1, 0, 0, 2, 2, 2, 0, 0, 2, 1, 2.00001f}, new double[]{0, 0, 0, 2, 2, 2, 2, 0});
+        testConvexHull_misdake(new double[]{0, 0, 1, 0, 0, 1, 1 / 3f, 2 / 3f, 2 / 3f, 1 / 3f}, new double[]{0, 0, 1, 0, 0, 1});
     }
 
-    private void testConvexHull_misdake(float[] in, float[] out) {
+    private void testConvexHull_misdake(double[] in, double[] out) {
         List<Point> points = new Polygon(in).points;
         Polygon polygon = ConvexHull.convexHull_misdake(points);
         List<Point> ref = new Polygon(out).points;
